requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-text-splitters
sentence-transformers
pypdf
docx2txt
unstructured
fastapi
uvicorn
streamlit
python-dotenv
pinecone-client
langchain-pinecone
langchain-groq
langchain-tavily
requests
uuid
langchain-huggingface
